moved vector functions to own helper
