+++ helpers/vector_database.js
@@ -1,0 +1,31 @@
+
+// Function for the search box
+async function vectorSearch(vectorArray){
+    // Perform the search with the vectorized input
+    const searchVector = [pgvector.toSql(vectorArray)];
+    const { rows } = await client.query(
+      `SELECT uuid, title, title_vector <-> $1 AS distance 
+         FROM text_embedding_3_large 
+         ORDER BY title_vector <-> $1 LIMIT 20`,
+      searchVector
+    );
+  
+    return rows;
+}
+
+// Search function used for RAG
+async function RagVectorSearch(vectorArray){
+    // Perform the search with the vectorized input
+    const searchVector = [pgvector.toSql(vectorArray)];
+    const { rows } = await client.query(
+      `SELECT uuid, title, abstract, image, title_vector <-> $1 AS distance 
+         FROM text_embedding_3_large 
+         ORDER BY title_vector <-> $1 LIMIT 3`,
+      searchVector
+    );
+  
+    return rows;
+}
+
+
+module.exports = { vectorSearch, RagVectorSearch };
